GridBagLayout added
I need to fix the buttons, they show up in the middle and they're too
big. They need to run across the top, and then another set will be
running along the bottom. The middle of the screen is left for the
swatch itself. Also, I need to make the screen non-resizable.