Fixing the oddCreateColorsMethod Swatch ERRORS
Moved pastelColors to Swatch. Added information to Creator. Changed
information in MonochromeSwatch's comments. oddCreateColors has much
less code in it now, still a bit repetitive, though. Consider a way to
combine the two into one...? Also, needs a sorting algorthm to complete
it. Test with 5 colors, should print 3 of either shades or tints and 2
of the opposite type. THIS CODE DOES NOT RUN W/O ERRORS.